Update player, shot, team, models & tests

+++ models/shot.go
@@ -8,7 +8,7 @@ type Shot struct {
 	ID        string  `db:"id" json:"id"`
 	Player    *Player `db:"player_id" json:"player,omitempty"`
 	Game      *Game   `db:"game_id" json:"game,omitempty"`
-	PtValue   uint8   `db:"pt_value" json:"pt_value"`
+	PtValue   uint64  `db:"pt_value" json:"pt_value"`
 	Made      bool    `db:"made" json:"made"`
 	XAxis     int     `db:"x_axis" json:"x_axis"`
 	YAxis     int     `db:"y_axis" json:"y_axis"`
